channels do ptr-int transmutes so move them to non-check-number-validity test
