+++ tests/run-pass/concurrency/channels.stderr
@@ -1,0 +1,2 @@
+warning: thread support is experimental and incomplete: weak memory effects are not emulated.
+
