+++ tests/run-pass/concurrency/simple.rs
@@ -1,4 +1,5 @@
 // ignore-windows: Concurrency on Windows is not supported yet.
+// compile-flags: -Zmiri-check-number-validity
 
 use std::thread;
