+++ tests/run-pass/concurrency/simple.stderr
@@ -1,5 +1,5 @@
 warning: thread support is experimental and incomplete: weak memory effects are not emulated.
 
-thread '<unnamed>' panicked at 'Hello!', $DIR/simple.rs:54:9
+thread '<unnamed>' panicked at 'Hello!', $DIR/simple.rs:55:9
 note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
-thread 'childthread' panicked at 'Hello, world!', $DIR/simple.rs:64:9
+thread 'childthread' panicked at 'Hello, world!', $DIR/simple.rs:65:9
